interfaces turned to types and api_callers renamed to middlewareHelpers
